fix: always output `.then(res => console.log(res))` in code sample
Since we dont know if the response is json or not, we can't make
assumptions. In an ideal world we'd conditionally do this based
on the accept header in the response, but Operation.getHeaders() only
returns with an array of headers and not their actual values. I think
this is good enough for now!

https://github.com/readmeio/api/pull/240#discussion_r569829932