feat: automatically parse the api response based on content-type  (#240)
* refactor: switch to using memfs instead of mock-fs

mock-fs has limitations which means you can use console.log during test runs
which is not ideal! Ended up doing something like this to get it to work:

https://github.com/tschaub/mock-fs/issues/234#issuecomment-653529125

* feat: automatically parse the api response based on content-type

This allows us to remove the `.then(res => res.json())` line from our
code samples.

```
const sdk = require('api')('@readme/5f7ceacf43621b0311080a58');

sdk.getAPISpecification({perPage: '10', page: '1'})
  .then(res => console.log(res))
  .catch(err => console.error(err));
```

We only perform the parsing if it's a successful response. If there's an error
case then we return with the original response object so you can access
the status as well as the actual body.

The code and tests for this have been stolen verbatim from the explorer:
https://github.com/readmeio/api-explorer/blob/77b90ebed4673f168354cdcd730e34b7ee016360/packages/api-explorer/src/lib/parse-response.js#L13-L30

BREAKING CHANGE: this is a breaking change.

* chore: relax commitlint rules on body and footer length

Taken from main codebase

* feat: remove res.json() line from the httpsnippet client

* fix: always output `.then(res => console.log(res))` in code sample

Since we dont know if the response is json or not, we can't make
assumptions. In an ideal world we'd conditionally do this based
on the accept header in the response, but Operation.getHeaders() only
returns with an array of headers and not their actual values. I think
this is good enough for now!

https://github.com/readmeio/api/pull/240#discussion_r569829932