fix(auth): remove the ability to have auth chained with each request
There's [a bug](https://github.com/readmeio/api/issues/102) with chaining
auth calls due to the fact that we have a global state, and calling
auth() again appends on to the state and breaks prepareAuth.

This PR removes the ability to call auth() in a chainable fashion, and
only allows auth to be called at the top level.

I'd like to figure out a way to bring this back eventually, but the
current API design doesn't support it in a clean way. We can revisit
this when we look to support operations with multiple auth mechanisms.

Fixes https://github.com/readmeio/api/issues/102