feat(api): export SDK
In some use cases, the SDK class needs to be instantiated several
times.

For instance, in a multi-tenancy scenario, there may be multiple
instances of the same API server, but serving different data from
different base URLs. In this case, we need to instantiate multiple
clients with different server URLs.